feat : stat date asset 생성 및 타격 피격 대기 애니메이션 추가

+++ Source/ProjectRAS/Character/Monster/Common/RASCommonMonster.h
@@ -25,7 +25,7 @@ public:
 
 	virtual void EndAttack() override;
 
-	virtual void HitFromActor(class ARASCharacterBase* InFrom, int InDamage) override;
+	virtual void HitFromActor(class ARASCharacterBase* InFrom, float InDamage) override;
 
 	virtual void KnockbackToDirection(class AActor* InFrom, FVector Direction, float InPower) override;
 protected:
